backend/API/requirements.txt
absl-py==0.8.1
astor==0.8.0
beautifulsoup4==4.8.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
fake-useragent==0.1.11
Flask==1.1.1
gast==0.2.2
google-api-core==1.14.3
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-cloud-language==1.3.0
google-pasta==0.1.8
googleapis-common-protos==1.6.0
grpcio==1.25.0
gunicorn==20.0.0
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.3
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.2
six==1.13.0
soupsieve==1.9.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tldextract==2.2.2
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
